Updated basket classes

+++ src/main/java/com/example/Book_Store/service/implementation/BookServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.Book_Store.service.implementation;
 import com.example.Book_Store.entities.Book;
 import com.example.Book_Store.repository.BookRepository;
 import com.example.Book_Store.service.BookService;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,7 +11,7 @@ import java.util.Optional;
 
 @Service
 public class BookServiceImpl implements BookService {
-   private final BookRepository bookRepository;
+    private final BookRepository bookRepository;
 
     public BookServiceImpl(BookRepository bookRepository) {
         this.bookRepository = bookRepository;
@@ -18,7 +19,8 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Book findBookById(Integer id) {
-        return bookRepository.findBookById(id);
+        return bookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Book not found with id: " + id));
     }
 
     @Override
@@ -28,31 +30,44 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public List<Book> findByCategoryName(String name) {
-        return bookRepository.findByCategoryName(name);
+        List<Book> books = bookRepository.findByCategoryName(name);
+        if (!books.isEmpty()) {
+            throw new ResourceNotFoundException("book list is empty");
+        }
+        return books;
     }
 
     @Override
     public List<Book> findByCategoryId(Integer id) {
-        return bookRepository.findByCategoryId(id);
-    }
-
-    @Override
-    public List<Book> findByPartialTitle(String partialTitle) {
-        return bookRepository.findByPartialTitle(partialTitle);
+        List<Book> books = bookRepository.findByCategoryId(id);
+        if (!books.isEmpty()) {
+            throw new ResourceNotFoundException("book list is empty");
+        }
+        return books;
     }
 
     @Override
     public void deleteBookById(Integer id) {
-    bookRepository.deleteBookById(id);
+
+        Book bookToDelete = bookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Book not found with id: " + id));
+        bookRepository.delete(bookToDelete);
     }
 
     @Override
     public void deleteBookByTitle(String title) {
-bookRepository.deleteBookByTitle(title);
+        if (!existsBookByTitle(title)) {
+            throw new ResourceNotFoundException("book not found");
+        }
+        bookRepository.deleteBookByTitle(title);
     }
 
     @Override
     public void deleteAllBooks() {
+        List<Book> books = findAllBooks();
+        if (books.isEmpty()) {
+            throw new ResourceNotFoundException("Books list is empty");
+        }
         bookRepository.deleteAll();
     }
 
@@ -68,16 +83,27 @@ bookRepository.deleteBookByTitle(title);
 
     @Override
     public Book createBook(Book book) {
+
         return bookRepository.save(book);
     }
 
     @Override
     public List<Book> findAllBooks() {
+        List<Book> books = bookRepository.findAll();
+        if (books.isEmpty()) {
+            throw new ResourceNotFoundException("Book list is empty");
+        }
         return bookRepository.findAll();
     }
 
     @Override
-    public Book updateBook(Book book) {
-        return bookRepository.save(book);
+    public Book updateBook(Integer id, Book book) {
+        Book existingBook = Optional.ofNullable(findBookById(id)).orElseThrow(() -> new ResourceNotFoundException("Book not found"));
+        existingBook.setCategory(book.getCategory());
+        existingBook.setTitle(book.getTitle());
+        existingBook.setAuthor(book.getAuthor());
+        existingBook.setPrice(book.getPrice());
+        existingBook.setQuantity(book.getQuantity());
+        return bookRepository.save(existingBook);
     }
 }
